+++ packages/expect/src/assertions/index.ts
@@ -1,0 +1,5 @@
+import toMatchObject from './toMatchObject'
+
+export default [
+  toMatchObject,
+]
